Extract simplification of token stream
Pull the token stream transformation out of the test case into its own component.
